pyproject.toml
[tool.poetry]
name = "budget-assistant"
version = "0.1.0"
description = "Smart Budget Assistant with AI-powered insights"
authors = ["Budget Team <team@budget.com>"]

[tool.poetry.dependencies]
python = ">=3.11.0,<3.13"
fastapi = ">=0.116.0"
uvicorn = ">=0.27.0"
python-multipart = ">=0.0.6"
aiofiles = ">=24.1.0"
jinja2 = ">=3.1.2"
httpx = ">=0.26.0"
python-dotenv = ">=1.0.0"
passlib = {extras = ["bcrypt"], version = ">=1.7.4"}
python-jose = {extras = ["cryptography"], version = ">=3.3.0"}
email-validator = ">=2.2.0"
openai = ">=1.98.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"